refactor dpm campaign workflow board routes
