+++ src/app/router_registry.py
@@ -55,6 +55,9 @@ from app.routers.dpm_wave_campaign_readiness import (
 from app.routers.dpm_wave_campaign_workflow import (
     router as dpm_wave_campaign_workflow_router,
 )
+from app.routers.dpm_wave_campaign_workflow_boards import (
+    router as dpm_wave_campaign_workflow_boards_router,
+)
 from app.routers.dpm_wave_evidence import router as dpm_wave_evidence_router
 from app.routers.dpm_waves import router as dpm_waves_router
 from app.routers.foundation import router as foundation_router
@@ -125,6 +128,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_portfolio_memory_router)
     app.include_router(dpm_wave_campaign_definitions_router)
     app.include_router(dpm_wave_campaign_discovery_router)
+    app.include_router(dpm_wave_campaign_workflow_boards_router)
     app.include_router(dpm_wave_campaign_workflow_router)
     app.include_router(dpm_construction_router)
     app.include_router(dpm_proof_packs_router)
